+++ src/main/java/com/porfolio/PRM/Repository/RUser.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RUser extends JpaRepository<EUser, Integer>{ 
+public interface RUser extends JpaRepository<EUser, Long>{ 
     public Optional<EUser> findByNombre(String nombre);
     public boolean existsByNombre(String nombre);
 }
